requirements.txt
wheel
absl-py==0.10.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
joblib
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lz4==3.1.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.19.5
psutil==5.7.2
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==1.2.1
scipy==1.5.2
#six==1.15.0
#smmap==3.0.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.50.2
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
shap
seaborn
ipython
ipykernel